% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cbindList}
\alias{cbindList}
\title{cbind a set of dataframes}
\usage{
cbindList(objs, dup.rm = T)
}
\arguments{
\item{objs}{character vector, the names of dataframes}

\item{dup.rm}{if `TRUE`, the variables with the same name will be delete and just the most left side variable will be retained\n
if `FALSE`, all the variables will be retained but the variables with the same name will be renamed}
}
\description{
improved version of rbind, it can rbind a set of dataframes which have the same colomns
}
\examples{
data(iris)
x1 <- iris
x2 <- iris
x3 <- iris
names(cbindList(c('x1', 'x2', 'x3'), dup.rm = T))
names(cbindList(c('x1', 'x2', 'x3'), dup.rm = F))
}

