% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Readfile.r
\name{data_proc}
\alias{data_proc}
\title{Manipulates the data in various ways}
\usage{
data_proc(wmean = TRUE, startle = TRUE, fulldays = TRUE, path = FALSE)
}
\description{
Manipulates the data in various ways
}
