% Generated by roxygen2 (4.0.2): do not edit by hand
\name{duplicate}
\alias{duplicate}
\title{Determine Duplicate Rows}
\usage{
duplicate(data, by)
}
\arguments{
\item{data}{data.frame or data.table}

\item{by}{character vector, `by` variables' names}
}
\description{
An improved version of `data.table::duplicated`. It returns all the rows which have duplicate rows, which is the difference from `duplicated`
}
\examples{
data(iris)
duplicate(iris, c('Sepal.Width', 'Species'))
}

