\name{getMzs}
\alias{getMzs}
\title{Extract M/Z values from the biomarker dataframe. }
\description{
Turn column names of the biomarker dataframe to numeric
M/Z values.
}
\usage{
getMzs(df)
}
\arguments{
  \item{df}{The biomarker dataframe with rows as spectra and columns as biomarkers.}
}
\value{
A numeric vector.
}
\author{Xiaochun Li}
\keyword{function}
