% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lof}
\alias{lof}
\title{An implementation of the LOF algorithm in parallel computing method}
\usage{
lof(data, k, equal.num = 0)
}
\arguments{
\item{data}{matrix or a data set that will be internally coerced into a matrix.}

\item{equal.num}{the max number of neighbours whose distance equals k_distance. `equal.num` = 0, by default}

\item{The}{number of neighbours that will be used in the calculation of the local outlier factors.}
}
\description{
This function obtain local outlier factors using the LOF algorithm. Namely, given a data set it produces a vector of local outlier factors for each case.
}

