% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlink.r
\name{mlink}
\alias{mlink}
\title{links the metadata to the experiment data (CAUTION: VERY SLOW)}
\usage{
mlink(metadata, processed = TRUE, bound = TRUE)
}
\description{
links the metadata to the experiment data (CAUTION: VERY SLOW)
}
