% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rbindList}
\alias{rbindList}
\title{combine R objects by rows}
\usage{
rbindList(objs)
}
\arguments{
\item{objs}{character vector, the names of dataframes}
}
\description{
improved version of rbind
}
\examples{
data(iris)
x1 <- iris
x2 <- iris
x3 <- iris
rbindList(c('x1', 'x2', 'x3'))
}

