% Generated by roxygen2 (4.0.2): do not edit by hand
\name{recode}
\alias{recode}
\title{An improved version of `car::recode`}
\usage{
recode(var, recodes, as.factor.result = TRUE, levels, labels)
}
\arguments{
\item{var}{numeric vector, character vector, factor, time vector}

\item{recodes}{character string of recode specification}

\item{as.factor.result}{if `TRUE`, it returns a factor}

\item{levels,}{labels Only as `as.factor.result` equals TRUE, then `levels` and `labels` work}
}
\description{
Add a function of recoding time vector to `car::recode`
}

