% Generated by roxygen2 (4.0.2): do not edit by hand
\name{round}
\alias{round}
\title{the improved version of base::round function}
\usage{
round(x, unit)
}
\arguments{
\item{x}{numeric vectors which will be rounded}

\item{unit}{the unit which is smaller than x}
}
\description{
get the number closest to its first argument which can be divided by exactly unit
}
\examples{
round(12345, 100)
round(1.098, 0.03)
}

