\name{SpecGrp}
\alias{SpecGrp}
\docType{data}
\title{A dataframe of spectrum ids and corresponding phenotype groups.}
\description{
A dataframe of spectrum id and corresponding phenotype group with respect to breast cancer.
}
\usage{data(SpecGrp)}
\format{
  A data frame with 167 observations on the following 2 variables.
  \describe{
    \item{GROUP}{a character vector with four possible values, "A", "B", "C" and "D".  "A" represents HER2 positive patients, "B" normal healthy patients, "C" ER/PR positive patients and "D" samples from a single healthy woman.}
    \item{SPECTRUM}{a character vector, the spectrum id of each sample.}
  }
}
\source{
Alex Miron's Lab at the Dana-Farber Cancer Institute.
}
\keyword{datasets}
